# contracts/core-business.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus Core Business Service API
  version: 1.0.0
  description: |
    API Contract-First của dịch vụ xử lý nghiệp vụ trung tâm (Nhóm B6 - Product B).
    Hợp đồng này quy định luồng giao tiếp REST đồng bộ với Access Gate Service (Nhóm B3)
    và các quy chuẩn quản lý cảnh báo vận hành trong Campus.
  contact:
    name: Đỗ Quang Minh (Group 6B Lead)
    email: 1771020481@dnu.edu.vn

servers:
  - url: http://localhost:4010
    description: Prism Mock Server dùng cho Lab 02

security:
  - bearerAuth: []

# Đăng ký danh mục thẻ toàn cục để sửa triệt để lỗi operation-tag-defined
tags:
  - name: Hệ thống
    description: Các endpoint kiểm tra trạng thái và vận hành hệ thống hạ tầng.
  - name: Giám sát Ra Vào
    description: Tiếp nhận dữ liệu đồng bộ từ Access Gate Service phục vụ kiểm soát an ninh.
  - name: Quản lý Cảnh báo
    description: Truy vấn và xử lý danh sách các cảnh báo (Alerts) nghiệp vụ trung tâm.

paths:
  /health:
    get:
      summary: Kiểm tra trạng thái hoạt động của dịch vụ
      description: Kiểm tra xem Core Business Service có đang sẵn sàng nhận sự kiện và kết nối DB/Queue không.
      operationId: checkHealth
      tags:
        - Hệ thống
      security: [] # Endpoint công khai không cần token
      responses:
        '200':
          description: Dịch vụ hoạt động bình thường
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - timestamp
                properties:
                  status:
                    type: string
                    examples: ["UP"]
                  timestamp:
                    type: string
                    format: date-time
                    examples: ["2026-05-18T10:30:00Z"]

  /api/v1/events/access:
    post:
      summary: Tiếp nhận sự kiện quẹt thẻ từ Access Gate Service
      description: Nhóm B3 gọi endpoint này đồng bộ để Core kiểm tra luật và quyết định đóng/mở cổng vật lý tại chỗ.
      operationId: processAccessEvent
      tags:
        - Giám sát Ra Vào
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericCampusEvent'
      responses:
        '201':
          description: Ghi nhận sự kiện thành công và trả về quyết định đóng/mở
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessEventResponse'
        '400':
          description: Dữ liệu gửi sang sai cấu trúc schema
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Thiếu mã xác thực hoặc mã không hợp lệ
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Vi phạm quy tắc dữ liệu hoặc xung đột logic hệ thống
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      summary: Lấy lịch sử sự kiện quẹt thẻ
      description: Truy vấn danh sách các sự kiện quẹt thẻ đã xảy ra tại các cổng an ninh. Hỗ trợ phân trang bằng cursor.
      operationId: listAccessEvents
      tags:
        - Giám sát Ra Vào
      parameters:
        - name: limit
          in: query
          required: false
          description: Số lượng bản ghi tối đa trả về trên một trang.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            examples: [20]
        - name: cursor
          in: query
          required: false
          description: Con trỏ phân trang (cursor) nhận từ phản hồi trước để truy vấn trang kế tiếp.
          schema:
            type: string
            examples: ["eyJpZCI6IkFDQy0yMDI2MDUxOC0wMDA1In0="]
      responses:
        '200':
          description: Truy vấn lịch sử quẹt thẻ thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - data
                  - pagination
                properties:
                  status:
                    type: string
                    examples: ["success"]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccessEventHistoryItem'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '401':
          description: Không có quyền truy cập
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/v1/alerts:
    get:
      summary: Lấy danh sách các cảnh báo vận hành hiện tại
      description: Lấy toàn bộ danh sách các alert (an ninh, môi trường) đã được Core xử lý sinh ra. Hỗ trợ phân trang bằng cursor.
      operationId: listAlerts
      tags:
        - Quản lý Cảnh báo
      parameters:
        - name: limit
          in: query
          required: false
          description: Số lượng bản ghi tối đa trả về trên một trang.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            examples: [20]
        - name: cursor
          in: query
          required: false
          description: Con trỏ phân trang (cursor) nhận từ phản hồi trước để truy vấn trang kế tiếp.
          schema:
            type: string
            examples: ["eyJpZCI6IkFMVC1CNi0yMDI2MDUxOC0wMTAifQ=="]
      responses:
        '200':
          description: Truy vấn danh sách cảnh báo thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - data
                  - pagination
                properties:
                  status:
                    type: string
                    examples: ["success"]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
        '401':
          description: Không có quyền truy cập
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/v1/alerts/{id}:
    get:
      summary: Lấy chi tiết thông tin của một cảnh báo cụ thể
      description: Xem chi tiết nội dung, mức độ nghiêm trọng và trạng thái xử lý của một alert theo mã ID.
      operationId: getAlertById
      tags:
        - Quản lý Cảnh báo
      parameters:
        - name: id
          in: path
          required: true
          description: Mã định danh duy nhất của Alert cần truy vấn
          schema:
            type: string
            examples: ["ALT-B6-20260518-001"]
      responses:
        '200':
          description: Tìm thấy thông tin chi tiết cảnh báo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          description: Không tìm thấy mã Alert yêu cầu
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

webhooks:
  securityAlertWebhook:
    post:
      summary: Gửi cảnh báo an ninh khẩn cấp đến hệ thống đăng ký nhận webhook
      description: |
        Được kích hoạt tự động bởi Core Business Service khi phát hiện cảnh báo ở mức độ cao (high severity),
        nhằm đẩy thông tin sự cố tới các hệ thống giám sát của bên thứ ba.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_type
                - alert
                - timestamp
              properties:
                event_type:
                  type: string
                  examples: ["security.alert.dispatched"]
                alert:
                  $ref: '#/components/schemas/Alert'
                timestamp:
                  type: string
                  format: date-time
                  examples: ["2026-05-18T08:30:15Z"]
      responses:
        '200':
          description: Webhook được nhận và xử lý thành công từ phía máy chủ nhận
        '202':
          description: Webhook được tiếp nhận để xử lý bất đồng bộ

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Truyền mã JWT Token hợp lệ vào Header dưới dạng `Authorization: Bearer <token>`"

  schemas:
    # 1. Schema yêu cầu đẩy dữ liệu quẹt thẻ từ nhóm B3 sang
    AccessEventRequest:
      type: object
      required:
        - card_id
        - gate_id
        - direction
        - timestamp
      properties:
        card_id:
          type: string
          description: Mã UID của thẻ RFID thu được từ đầu đọc thẻ vật lý
          examples: ["RFID-2026-9999"]
        gate_id:
          type: string
          description: Mã định danh của cổng kiểm soát diễn ra hành vi quẹt thẻ
          examples: ["gate-lib-01"]
        direction:
          type: string
          description: Hướng di chuyển qua cổng của đối tượng
          enum: [IN, OUT]
          examples: ["IN"]
        timestamp:
          type: string
          format: date-time
          description: Thời gian quẹt thẻ theo chuẩn chuỗi ISO 8601 chốt từ đàm phán
          examples: ["2026-05-18T08:30:00Z"]

    # 2. Schema Core phản hồi đồng bộ trả lại cho nhóm B3 quyết định đóng/mở tại chỗ
    AccessEventResponse:
      type: object
      required:
        - access_granted
        - reason
        - person_id
      properties:
        access_granted:
          type: boolean
          description: Quyết định của bộ máy xử lý luật Core cho phép mở barie hay không
          examples: [true]
        reason:
          type: string
          description: Lý do cụ thể đi kèm quyết định phê duyệt/từ chối
          examples: ["Thẻ sinh viên hợp lệ và còn thời hạn truy cập"]
        person_id:
          type: string
          description: Mã số sinh viên hoặc nhân viên tương ứng với chủ thẻ
          examples: ["SV00123"]

    # 3. Schema mô tả một Cảnh báo (Alert)
    Alert:
      type: object
      required:
        - alert_id
        - type
        - severity
        - message
        - created_at
        - resolved_at
      properties:
        alert_id:
          type: string
          examples: ["ALT-B6-20260518-001"]
        type:
          type: string
          examples: ["suspicious_card_activity"]
        severity:
          type: string
          enum: [low, medium, high]
          examples: ["medium"]
        message:
          type: string
          examples: ["Cảnh báo: Thẻ RFID-2026-9999 được quẹt liên tiếp tại 2 cổng khác nhau trong vòng dưới 30 giây."]
        created_at:
          type: string
          format: date-time
          examples: ["2026-05-18T08:30:15Z"]
        resolved_at:
          type: [string, "null"]
          format: date-time
          description: Thời điểm sự cố được xử lý xong. Dùng mảng kiểu dữ liệu thay thế hoàn toàn cho nullable cũ của OpenAPI 3.0.

    # 4. Schema cấu trúc báo lỗi chuẩn hóa Problem Details (RFC 7807) áp dụng cho lỗi 4xx/5xx
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          examples: ["https://smartcampus.dnu.edu.vn/probs/bad-request"]
        title:
          type: string
          examples: ["Invalid Request Payload"]
        status:
          type: integer
          examples: [400]
        detail:
          type: string
          examples: ["Trường dữ liệu 'card_id' bắt buộc phải truyền và không được để trống."]
        instance:
          type: [string, "null"]
          format: uri
          examples: ["/api/v1/events/access"]

    # 5. Schema Đa hình mô phỏng Hệ sinh thái Sự kiện (oneOf + discriminator nâng cao)
    GenericCampusEvent:
      type: object
      required:
        - event_type
        - timestamp
      properties:
        event_type:
          type: string
          description: Trường định danh dùng để phân tách loại gói tin sự kiện đẩy về hệ thống
        timestamp:
          type: string
          format: date-time
      oneOf:
        - $ref: '#/components/schemas/AccessEventRequest'
      discriminator:
        propertyName: event_type
        mapping:
          access_event: '#/components/schemas/AccessEventRequest'

    # 6. Schema thông tin phân trang dựa trên Cursor (Cursor-based Pagination)
    PaginationInfo:
      type: object
      required:
        - next_cursor
        - has_more
      properties:
        next_cursor:
          type: [string, "null"]
          description: Con trỏ trỏ đến trang tiếp theo. Nếu là null nghĩa là không còn trang tiếp theo.
          examples: ["eyJpZCI6IkFMVC1CNi0yMDI2MDUxOC0wMTAifQ=="]
        has_more:
          type: boolean
          description: Cho biết có còn dữ liệu ở trang tiếp theo hay không.
          examples: [true]

    # 7. Schema chi tiết một bản ghi lịch sử sự kiện quẹt thẻ
    AccessEventHistoryItem:
      type: object
      required:
        - event_id
        - card_id
        - gate_id
        - direction
        - timestamp
        - access_granted
        - reason
      properties:
        event_id:
          type: string
          description: Mã định danh sự kiện (UUID)
          examples: ["acc-uuid-2026-b6"]
        card_id:
          type: string
          description: Mã UID của thẻ RFID
          examples: ["RFID-2026-9999"]
        gate_id:
          type: string
          description: Mã cổng kiểm soát
          examples: ["gate-lib-01"]
        direction:
          type: string
          enum: [IN, OUT]
          examples: ["IN"]
        timestamp:
          type: string
          format: date-time
          description: Thời gian quẹt thẻ
          examples: ["2026-05-18T08:30:00Z"]
        access_granted:
          type: boolean
          description: Kết quả phê duyệt quyền truy cập
          examples: [true]
        reason:
          type: string
          description: Lý do đi kèm quyết định
          examples: ["Thẻ hợp lệ"]